corrected P_gac radius(depends now on pixelsize from DEM) merge flow_sheds is more efficient

+++ other_glacier.py
@@ -40,79 +40,36 @@ from skimage.feature import peak_local_max
 from shapely.geometry import Point,Polygon,MultiPolygon
 
 if __name__ == '__main__':
-    RGI_FILE ='/home/juliaeis/Dokumente/OGGM/work_dir/CentralEurope3000m+/outlines.shp'
-    base_dir = os.path.join(os.path.expanduser('/home/juliaeis/Dokumente/OGGM/work_dir/CentralEurope3000m+'))
+    RGI_FILE ='/home/juliaeis/Dokumente/OGGM/work_dir/Alaska/RGI50-01.10689/outlines.shp'
+    base_dir = os.path.join(os.path.expanduser('/home/juliaeis/Dokumente/OGGM/work_dir/Alaska'))
     import time
     start0=time.time()
     cfg.initialize()
 
+    cfg.PATHS['dem_file']= os.path.join('/home/juliaeis/Dokumente/OGGM/work_dir/Alaska/dem.tif')
     # Read RGI file
     rgidf = salem.read_shapefile(RGI_FILE, cached=True)
-    f = open(base_dir + '/failed_glacier_id.txt', 'w')
-    g = open(base_dir + '/succes_glacier_id.txt', 'w')
+
     for i in range(len(rgidf)):
-        #if rgidf.iloc[i].RGIId not in ['RGI50-11.00116','RGI50-11.00181']:
-        gdir=oggm.GlacierDirectory(rgidf.iloc[i], base_dir=base_dir)
-        print gdir.dir
-        if gdir.has_file('outlines', div_id=0) and gdir.has_file('dem', div_id=0):
-            #tasks.define_glacier_region(gdir, entity=rgidf.iloc[i])
-            pass
-        else:
-            tasks.define_glacier_region(gdir, entity=rgidf.iloc[i])
-        input_shp = gdir.get_filepath('outlines', div_id=0)
-        input_dem = gdir.get_filepath('dem', div_id=0)
-
-        if os.path.isfile(gdir.dir+'/divide_01/outlines.shp'):
-            os.remove(gdir.dir+'/divide_01/outlines.shp')
-        try:
-            no=dividing_glaciers(input_dem, input_shp)
-            if no is not 1:
-                g.write(str(rgidf.iloc[i].RGIId) + '/n')
-        except:
-            f.write(str(rgidf.iloc[i].RGIId)+'/n' )
-    f.closed
-    g.closed
-
-    '''
-    base_dir = os.path.join(os.path.expanduser('/home/juliaeis/Dokumente/OGGM/work_dir'), 'Scandinavia')
-    #entity = gpd.GeoDataFrame.from_file(get_demo_file('Hintereisferner.shp')).iloc[0]
-    #gdir = oggm.GlacierDirectory(entity, base_dir=base_dir)
-    for dir in os.listdir(base_dir):
-        if dir.startswith('RGI50'):
-            gdir=oggm.GlacierDirectory(dir,base_dir=base_dir)
-            entity = gpd.GeoDataFrame.from_file(gdir.dir+'/outlines.shp').iloc[0]
+        if rgidf.iloc[i].RGIId in ['RGI50-01.10689']:
+            gdir=oggm.GlacierDirectory(rgidf.iloc[i], base_dir=base_dir)
             print gdir.dir
-            #check if required files exists
-            if gdir.has_file('outlines',div_id=0) and gdir.has_file('dem', div_id=0):
+            #tasks.define_glacier_region(gdir, entity=rgidf.iloc[i])
+
+            if gdir.has_file('outlines', div_id=0) and gdir.has_file('dem', div_id=0):
+                #tasks.define_glacier_region(gdir, entity=rgidf.iloc[i])
                 pass
             else:
-                tasks.define_glacier_region(gdir,entity=entity)
-            #tasks.glacier_masks(gdir)
-            ###################preprocessing########################
-            input_shp =gdir.get_filepath('outlines',div_id=0)
-            input_dem=gdir.get_filepath('dem',div_id=0)
+                tasks.define_glacier_region(gdir, entity=rgidf.iloc[i])
+            input_shp = gdir.get_filepath('outlines', div_id=0)
+            input_dem = gdir.get_filepath('dem', div_id=0)
 
-            #get pixel size
-            with rasterio.open(input_dem) as dem:
-                global pixelsize
-                pixelsize=int(dem.transform[1])
-            dividing_glaciers(input_dem, input_shp)
+            if os.path.isfile(gdir.dir+'/divide_01/outlines.shp'):
+                os.remove(gdir.dir+'/divide_01/outlines.shp')
 
-
-    print time.time()-start0
-
-    #test if it works
-
-    from oggm import graphics
-
-    tasks.glacier_masks(gdir)
-    tasks.compute_centerlines(gdir)
-    tasks.compute_downstream_lines(gdir)
-    tasks.catchment_area(gdir)
-    tasks.initialize_flowlines(gdir)
-    tasks.catchment_width_geom(gdir)
-
-    graphics.plot_centerlines(gdir)
-    plt.show()
-
-'''
+            no,k=dividing_glaciers(input_dem, input_shp)
+            '''
+            tasks.glacier_masks(gdir)
+            tasks.compute_centerlines(gdir)
+            graphics.plot_centerlines(gdir)
+            '''
